Honestly this is a really good base for the project!!
There are a few places code could be simplified and/or made into a method and re-used but no huge/glaring issues with the code itself.

I left comments in there for places to work on as far as making your code more readable/effeicient/correct. Text me about questions you have on the comments I left.